Use correct `defaults` sourcing for platform
There are two issues breaking this functionality.

- The functions.sh is currently sourcing only not platform specific
defaults file. This works sometimes but sometimes but it might went
wrong on the other than Fedora platforms.

- The PLATFORM_NAME is incorrectly propagated to tests as PLATFORM
variable. It was propagated from the container call this way and used
correctly in the run-kstest with this name. However, all these variables
are also propagated to shell scripts for given tests. There this name
fails because PLATFORM_NAME is used everywhere else.