Update pre-install test for new modularization
Due to modularization it was necessary to make a change in the tests because the triggered error from the module shows
a traceback in the log, which automatically terminates the test at --erroronfail. So the temporary solution is to test
the test without this parameter, and look for the error in the log.