Add AppStream repository for FTP access
The AppStream repository is required only for installation of RHEL and
not used on Fedora. To resolve that let's add a new fragment which will
be platform specific and set FTP as installation source.

The FTP AppStream source is required for a new flatpak tests which are
coming in later commit.