Add flatpak preinstall.d tests
For HTTP and FTP tests. Ideally we would be testing also NFS but we
are facing issues with NFS installation sources.

This is testing support of
https://github.com/rhinstaller/anaconda/pull/6440

This is unfortunately not reasonably possible to test this on Fedora as
the repositories needs to be created correctly and this is not supported
on Fedora yet.