Enable virt-viewer connect without root
The original reason why we have root, maybe not the only one, was that
podman under root is using bridging which add us possibility to connect
to this IP by virt-viewer to see the VM. This is must have for
reasonable debugging. However, it is not great from security point and
you are forced to deal with files created under root. Also the bridging
will make consuming VPN resources harder.

The default connection under user is using slirp4netns solution which
creates internal network stack. Which makes it harder to reach.

However, there is also solution to use 'host' network which is sharing
network with the host. That is considered unsecure because it will allow
access to DBus resources etc..., however, we are not providing server
and the container is under our control.

This solution will allow us to connect to 127.0.0.1 network.