Extract LUKS auto-unlock post script to reusable fragment
Create fragments/shared/storage/luks_auto_unlock_post.ks containing
the common LUKS auto-unlock setup code that was duplicated in
bootc-luks.ks.in and rpm-ostree-container-luks.ks.in.

The only difference is that in new file remounts boot as rw before
calling dracut. This is needed for bootc tests where /boot is RO at the
%post time.